Split test_process_pool: add test_thread_pool
