+++ Lib/test/test_concurrent_futures/test_process_pool.py
@@ -1,35 +1,17 @@
-import contextlib
-import multiprocessing as mp
-import multiprocessing.process
-import multiprocessing.util
 import os
 import sys
-import threading
 import time
 import unittest
-import weakref
 from concurrent import futures
 from concurrent.futures.process import BrokenProcessPool
 
 from test import support
 from test.support import hashlib_helper
 
+from .executor import ExecutorTest, mul
 from .util import (
-    BaseTestCase, ThreadPoolMixin, ProcessPoolForkMixin,
-    ProcessPoolForkserverMixin, ProcessPoolSpawnMixin,
+    ProcessPoolForkMixin, ProcessPoolForkserverMixin, ProcessPoolSpawnMixin,
     create_executor_tests, setup_module)
-
-
-def mul(x, y):
-    return x * y
-
-def capture(*args, **kwargs):
-    return args, kwargs
-
-
-class MyObject(object):
-    def my_method(self):
-        pass
 
 
 class EventfulGCObj():
@@ -38,175 +20,6 @@ class EventfulGCObj():
 
     def __del__(self):
         self.event.set()
-
-
-def make_dummy_object(_):
-    return MyObject()
-
-
-class ExecutorTest:
-    # Executor.shutdown() and context manager usage is tested by
-    # ExecutorShutdownTest.
-    def test_submit(self):
-        future = self.executor.submit(pow, 2, 8)
-        self.assertEqual(256, future.result())
-
-    def test_submit_keyword(self):
-        future = self.executor.submit(mul, 2, y=8)
-        self.assertEqual(16, future.result())
-        future = self.executor.submit(capture, 1, self=2, fn=3)
-        self.assertEqual(future.result(), ((1,), {'self': 2, 'fn': 3}))
-        with self.assertRaises(TypeError):
-            self.executor.submit(fn=capture, arg=1)
-        with self.assertRaises(TypeError):
-            self.executor.submit(arg=1)
-
-    def test_map(self):
-        self.assertEqual(
-                list(self.executor.map(pow, range(10), range(10))),
-                list(map(pow, range(10), range(10))))
-
-        self.assertEqual(
-                list(self.executor.map(pow, range(10), range(10), chunksize=3)),
-                list(map(pow, range(10), range(10))))
-
-    def test_map_exception(self):
-        i = self.executor.map(divmod, [1, 1, 1, 1], [2, 3, 0, 5])
-        self.assertEqual(i.__next__(), (0, 1))
-        self.assertEqual(i.__next__(), (0, 1))
-        self.assertRaises(ZeroDivisionError, i.__next__)
-
-    def test_map_timeout(self):
-        results = []
-        try:
-            for i in self.executor.map(time.sleep,
-                                       [0, 0, 6],
-                                       timeout=5):
-                results.append(i)
-        except futures.TimeoutError:
-            pass
-        else:
-            self.fail('expected TimeoutError')
-
-        self.assertEqual([None, None], results)
-
-    def test_shutdown_race_issue12456(self):
-        # Issue #12456: race condition at shutdown where trying to post a
-        # sentinel in the call queue blocks (the queue is full while processes
-        # have exited).
-        self.executor.map(str, [2] * (self.worker_count + 1))
-        self.executor.shutdown()
-
-    @support.cpython_only
-    def test_no_stale_references(self):
-        # Issue #16284: check that the executors don't unnecessarily hang onto
-        # references.
-        my_object = MyObject()
-        my_object_collected = threading.Event()
-        my_object_callback = weakref.ref(
-            my_object, lambda obj: my_object_collected.set())
-        # Deliberately discarding the future.
-        self.executor.submit(my_object.my_method)
-        del my_object
-
-        collected = my_object_collected.wait(timeout=support.SHORT_TIMEOUT)
-        self.assertTrue(collected,
-                        "Stale reference not collected within timeout.")
-
-    def test_max_workers_negative(self):
-        for number in (0, -1):
-            with self.assertRaisesRegex(ValueError,
-                                        "max_workers must be greater "
-                                        "than 0"):
-                self.executor_type(max_workers=number)
-
-    def test_free_reference(self):
-        # Issue #14406: Result iterator should not keep an internal
-        # reference to result objects.
-        for obj in self.executor.map(make_dummy_object, range(10)):
-            wr = weakref.ref(obj)
-            del obj
-            support.gc_collect()  # For PyPy or other GCs.
-            self.assertIsNone(wr())
-
-
-class ThreadPoolExecutorTest(ThreadPoolMixin, ExecutorTest, BaseTestCase):
-    def test_map_submits_without_iteration(self):
-        """Tests verifying issue 11777."""
-        finished = []
-        def record_finished(n):
-            finished.append(n)
-
-        self.executor.map(record_finished, range(10))
-        self.executor.shutdown(wait=True)
-        self.assertCountEqual(finished, range(10))
-
-    def test_default_workers(self):
-        executor = self.executor_type()
-        expected = min(32, (os.cpu_count() or 1) + 4)
-        self.assertEqual(executor._max_workers, expected)
-
-    def test_saturation(self):
-        executor = self.executor_type(4)
-        def acquire_lock(lock):
-            lock.acquire()
-
-        sem = threading.Semaphore(0)
-        for i in range(15 * executor._max_workers):
-            executor.submit(acquire_lock, sem)
-        self.assertEqual(len(executor._threads), executor._max_workers)
-        for i in range(15 * executor._max_workers):
-            sem.release()
-        executor.shutdown(wait=True)
-
-    def test_idle_thread_reuse(self):
-        executor = self.executor_type()
-        executor.submit(mul, 21, 2).result()
-        executor.submit(mul, 6, 7).result()
-        executor.submit(mul, 3, 14).result()
-        self.assertEqual(len(executor._threads), 1)
-        executor.shutdown(wait=True)
-
-    @unittest.skipUnless(hasattr(os, 'register_at_fork'), 'need os.register_at_fork')
-    def test_hang_global_shutdown_lock(self):
-        # bpo-45021: _global_shutdown_lock should be reinitialized in the child
-        # process, otherwise it will never exit
-        def submit(pool):
-            pool.submit(submit, pool)
-
-        with futures.ThreadPoolExecutor(1) as pool:
-            pool.submit(submit, pool)
-
-            for _ in range(50):
-                with futures.ProcessPoolExecutor(1, mp_context=mp.get_context('fork')) as workers:
-                    workers.submit(tuple)
-
-    def test_executor_map_current_future_cancel(self):
-        stop_event = threading.Event()
-        log = []
-
-        def log_n_wait(ident):
-            log.append(f"{ident=} started")
-            try:
-                stop_event.wait()
-            finally:
-                log.append(f"{ident=} stopped")
-
-        with self.executor_type(max_workers=1) as pool:
-            # submit work to saturate the pool
-            fut = pool.submit(log_n_wait, ident="first")
-            try:
-                with contextlib.closing(
-                    pool.map(log_n_wait, ["second", "third"], timeout=0)
-                ) as gen:
-                    with self.assertRaises(TimeoutError):
-                        next(gen)
-            finally:
-                stop_event.set()
-            fut.result()
-        # ident='second' is cancelled as a result of raising a TimeoutError
-        # ident='third' is cancelled because it remained in the collection of futures
-        self.assertListEqual(log, ["ident='first' started", "ident='first' stopped"])
 
 
 class ProcessPoolExecutorTest(ExecutorTest):
